feat: translating static text to language file
This is in an effort to not only encourage more use through
locallization, but also to make the text easier to manage by being all
in one spot. This is only the start of a broader effort. Many of the
normalized fields in the database still need to be language filed (e.g.
Markets)